+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.2.0]
+- Update `@ramstack/hotkey` package
+- Support `trusted` modifier
+
 ## [1.1.0]
 - Add hotkey property to determinie which of hotkeys triggered the event
 - Add Vue plugin